fixed service worker
